% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab1sum.R
\name{lab1sum}
\alias{lab1sum}
\title{Lab 1 Summary Function}
\usage{
lab1sum(x)
}
\arguments{
\item{x}{a data set input}
}
\value{
a summary of variables in the data set
}
\description{
Lab 1 Summary Function
}
\examples{
lab1sum(x = c(1,2,3,4,5))
}
