% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2zvalue.R
\name{lab2zvalue}
\alias{lab2zvalue}
\title{Lab 2 Z Value}
\usage{
lab2zvalue(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a numerical vector
}
\description{
Lab 2 Z Value
}
\examples{
lab2class(x=data$column)
}
