% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3lmsummary.R
\name{lab3lmsummary}
\alias{lab3lmsummary}
\title{Lab 3 Linear Model Summary Coefficients}
\usage{
lab3lmsummary(x)
}
\arguments{
\item{x}{linear model input}
}
\value{
numerical vector output
}
\description{
Lab 3 Linear Model Summary Coefficients
}
\examples{
mydata.lm = NULL
lab3lmsummary(x=mydata.lm)
}
