% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{myboot}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{a numerical vector}

\item{x}{a data frame numerical vector}

\item{fun}{a defined function}

\item{alpha}{a numerical vector between 0 and 1}

\item{cx}{a numerical vector}

\item{...}{addtional parameters}
}
\value{
a list of values from the function including a histogram
}
\description{
myboot
}
\examples{
myboot(iter=10000,x = c(0:20),fun="mean",alpha=0.5)
}
