% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{numerical vector of sample size}

\item{iter}{numerical vector of iteration size}

\item{p}{numerical vector of probabality value}

\item{...}{numerical vectors}
}
\value{
graph
}
\description{
mycltb
}
\examples{
mycltb(n=4,iter=10000,p=0.3)
}
