% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of iterations}

\item{N}{number of data points}

\item{r}{number of successful points}

\item{n}{number of sample points}
}
\value{
a set of graphs
}
\description{
Hypergeometric Function
}
\examples{
myhyper(iter=1000, n=19, N=20, r=12)
}
