% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{a numeric vector, the value up to which area is found}

\item{mu}{a numeric vector, the mean value of the function}

\item{sigma}{a numeric vector, the standard deviation of the function}
}
\value{
a graph, and a list of values
}
\description{
myncurve
}
\examples{
myncurve(1,2,3)
}
