% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{a numeric vector}

\item{gamma}{a numeric vector}

\item{p}{a numeric vector}
}
\value{
a list of values
}
\description{
ntickets
}
\examples{
ntickets(N = 200, gamma = 0.02, p=0.95)
}
